+++ EquationElements/Number/Overload LessThanOrEqualTo.cs
@@ -4,13 +4,13 @@
     {
         public static bool operator <=(Number a, Number b) => a < b || a == b;
 
-        public static bool operator <=(Number a, int b) => a < b || a == b;
+        public static bool operator <=(Number a, long b) => a < b || a == b;
 
         public static bool operator <=(Number a, decimal b) => a < b || a == b;
 
         public static bool operator <=(Number a, double b) => a < b || a == b;
 
-        public static bool operator <=(int a, Number b) => a < b || a == b;
+        public static bool operator <=(long a, Number b) => a < b || a == b;
 
         public static bool operator <=(decimal a, Number b) => a < b || a == b;
