+++ EquationElements/Number/Overload Mod.cs
@@ -7,7 +7,7 @@ namespace EquationElements
                 ? new Number(a.AsDecimal % b.AsDecimal)
                 : new Number(a.AsDouble % b.AsDouble);
 
-        public static Number operator %(Number a, int b) =>
+        public static Number operator %(Number a, long b) =>
             a.IsDecimal
                 ? new Number(a.AsDecimal % b)
                 : new Number(a.AsDouble % b);
@@ -19,7 +19,7 @@ namespace EquationElements
 
         public static Number operator %(Number a, double b) => new Number(a.AsDouble % b);
 
-        public static Number operator %(int a, Number b) =>
+        public static Number operator %(long a, Number b) =>
             b.IsDecimal
                 ? new Number(a % b.AsDecimal)
                 : new Number(a % b.AsDouble);
